src/datahub/databus/databus-puller-datanode/src/main/java/com/tencent/bk/base/datahub/databus/connect/source/datanode/transform/TransformResult.java
/*
 * Tencent is pleased to support the open source community by making BK-BASE 蓝鲸基础平台 available.
 *
 * Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.
 *
 * BK-BASE 蓝鲸基础平台 is licensed under the MIT License.
 *
 * License for BK-BASE 蓝鲸基础平台:
 * --------------------------------------------------------------------
 * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
 * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
 * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
 * and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all copies or substantial
 * portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
 * LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
 * NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
 * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */

package com.tencent.bk.base.datahub.databus.connect.source.datanode.transform;

import org.apache.kafka.clients.consumer.ConsumerRecord;

import java.util.ArrayList;
import java.util.Iterator;
import java.util.List;

/**
 * 转换结果封装类
 */
public class TransformResult implements Iterable<TransformRecord> {

    private List<TransformRecord> transRecords = new ArrayList<>();

    public TransformResult() {
    }


    /**
     * 实现迭代器遍历所有的按bid 分组的record
     *
     * @return 按bid 分组的record
     */
    @Override
    public Iterator<TransformRecord> iterator() {
        return transRecords.iterator();
    }

    /**
     * 记录转换记录
     *
     * @param bid 业务id
     * @param record 业务id对应的record
     */
    public void putRecord(String bid, ConsumerRecord<byte[], byte[]> record) {
        transRecords.add(new TransformRecord(bid, record));
    }


}
